fix: checklist bag, column coping, deleting file

+++ server/routes/uploadRoute.js
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 import { unlink } from 'fs/promises';
 import upload from '../middlewares/upload.js';
+import Card from '../models/cardModel.js';
 
 const router = Router();
 
@@ -19,8 +20,12 @@ router.post('/', upload.single('file'), (req, res) => {
 router.post('/delete', async (req, res) => {
   try {
     const { path } = req.body;
-    await unlink(`.${path}`);
-    return res.status(200).json({ message: 'Файл удален' });
+    const cards = await Card.find({ 'attachments.url': path });
+    if (cards.length === 1) {
+      await unlink(`.${path}`);
+      return res.status(200).json({ message: 'Файл удален' });
+    }
+    return res.status(200).json({ message: 'Ссылка на файл используется в других карточках' });
   } catch (error) {
     return res.status(500).json({ error: error.message });
   }
